% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AdaptMetrop}
\alias{AdaptMetrop}
\title{Turn a non-adaptive Metropolis sampler into an adaptive Metropolis sampler}
\usage{
AdaptMetrop(f, s, batch.size = 50, target = 0.44, DeltaN)
}
\description{
Given a non-adpative sampler of the form f(..., s), \code{AdaptMetrop} will
return a function g(...) that automatically adapts the Metropolis proposal
standard deviation s to try and acheive a target acceptance rate.
}
\details{
\code{AdaptMetrop} uses the algorithm from Roberts & Rosenthal (2009), which
re-scales the proposal standard deviation after a fixed number of MCMC
iterations have elapsed.  The goal of the algorithm is to acheive a target
acceptance rate for the Metropolis step.  After the n\emph{th} batch of MCMC
iterations the proposal standard deviation, \eqn{log(s)}, is
increased/decreased by \eqn{\delta(n)}.  \eqn{log(s)} is increased by
\eqn{\delta(n)} if the observed acceptance rate is more than the target
acceptance rate, or decreased by \eqn{\delta(n)} if the observed acceptance
rate is less than the target acceptance rate.

\eqn{\delta(n)} is set to \eqn{\delta(n) = min(0.01, n^{-1/2})} unless
re-specified in the function call. Some care should be taken if re-specifying
\code{DeltaN}, as the ergodicity of the chain may not be preserved if
certain conditions aren't met.  See Roberts & Rosenthal (2009) in the
references for details.

\code{AdaptMetrop} keeps track of the the acceptance rate by comparing the
previously sampled value from \code{f} to the next value.  If the two values
are equal, the proposal is considered to be rejected, whereas if the two
values are different the proposal is considered accepted. \code{AdaptMetrop},
therefore, is intended to be used in cases where the target distribution is
continuous and ties aren't possible.  Otherwise the calculated acceptance
rate may be incorrect. For example, if the target distribution and the
proposal distribution are discrete, it may be possible for the previous
parameter value and the next parameter value to be the same, even if the
proposal is accepted.

The proposal standard deviation \code{s} can be either a vector or a scalar.
If the initial value of \code{s} is a scalar, \eqn{f} will be treated as a
sampler for a random vector or a joint parameter update.  Alternatively, if
the dimension of \eqn{s} is equal to the dimension of the parameters returned
by \eqn{f}, the individual elements \eqn{s} will be treated as individual
proposal standard deviations for the elements of \eqn{f}.  This functionality
can be used, for example, if \eqn{f} loops through each of its elements
individually, updating each using Metropolis step.  In this case, \eqn{f}
should contructed to receive \eqn{s} as a vector argument.
}
\examples{
# TODO
}
\references{
Gareth O. Roberts & Jeffrey S. Rosenthal (2009) Examples of
  Adaptive MCMC, Journal of Computational and Graphical Statistics, 18:2,
  349-367, DOI:
  \href{https://doi.org/10.1198/jcgs.2009.06134}{10.1198/jcgs.2009.06134}
}
