% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions.R
\name{get_division}
\alias{get_division}
\title{get_division}
\usage{
get_division(primary_name, type = NULL, as_sf = TRUE)
}
\arguments{
\item{primary_name}{the primary name of the division, e.g. "United States",
or "California".}

\item{type}{the Overture subtype, e.g. "country", "region", "county", "locality",
"neighborhood" etc. By default will search all types.}

\item{as_sf}{Return an \code{sf} object?  default \code{TRUE}. Set to FALSE only to continue
additional processing outside-of-RAM with duckdbfs spatial tools.}
}
\description{
Return a polygon for the desired division (e.g. country, region, county, etc).
By default will search for a direct match to the Overture primary name across
all Overture subtypes.  Specify the type for faster search and/or to resolve
ambiguities.
}
\examples{
\dontshow{if ( interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

uk <- get_division("United Kingdom")
ca <- get_division("California", type = "region")
library(sf)
plot(uk[1])
\dontshow{\}) # examplesIf}
}
