% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions.R
\name{get_subdivision}
\alias{get_subdivision}
\title{get_subdivision}
\usage{
get_subdivision(primary_name, type = c("country", "region", "county"))
}
\arguments{
\item{primary_name}{the name listed as "primary" in Overture.}

\item{type}{type of division: country, region, county}
}
\value{
an \code{sf} object whose rows are all the polygon (or multipolygon)
features that are immediate children of the matching division.
}
\description{
Searches the division theme for a matching primary name.  Returns all
polygons (division_area) for which the matching primary name is listed
as the parent_division_id.  For instance, \code{get_subdivision("United States")}
returns a data frame containing the 50 states and 7 territories in the
sovereign US., while \code{get_subdivision("California")} returns the counties
in California.
}
\details{
NOTE: This can be very slow without a very high-speed network connection.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

gdf <- get_subdivision("United States")
map(gdf)

gdf <- get_subdivision("California")
map(gdf)
\dontshow{\}) # examplesIf}
}
\references{
https://docs.overturemaps.org/schema/reference/divisions/division/
}
