% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgl.R
\name{map}
\alias{map}
\title{Convenience function to generate mapgl maps}
\usage{
map(gdf, tooltip = "primary", fill_color = "blue", fill_opacity = 0.1, ...)
}
\arguments{
\item{gdf}{an \code{sf} object, URL, or \code{tbl_duckdb_connection}
to a spatial object}

\item{tooltip}{column that should be used as the tooltip}

\item{fill_color}{polygon fill color}

\item{fill_opacity}{polygon fill opacity}

\item{...}{additional parameters to \code{\link[mapgl:add_fill_layer]{mapgl::add_fill_layer()}}}
}
\value{
a magl map object
}
\description{
Convenience function to generate mapgl maps
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ri <- get_division("Rhode Island", type="region")
map(ri)
\dontshow{\}) # examplesIf}
}
