% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overture.R
\name{overture}
\alias{overture}
\title{Access overture data}
\usage{
overture(theme = "divisions", type = "division_area")
}
\arguments{
\item{theme}{overture theme}

\item{type}{data type (overture partition, not 'subtype')}
}
\description{
Access overture data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
overture()

# example code
\dontshow{\}) # examplesIf}
}
