% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{Resume}
\alias{Resume}
\title{Resumes an interrupted file-backed MCMC}
\usage{
Resume(backing.path)
}
\arguments{
\item{backing.path}{directory path where MCMC samples were saved}
}
\description{
Resumes an interrupted file-backed MCMC
}
