% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgl.R
\name{to_s3}
\alias{to_s3}
\title{to_s3}
\usage{
to_s3(data, path, id_col = "id")
}
\arguments{
\item{data}{gdf or duckdbfs table connection (e.g. from \code{overture()})}

\item{path}{bucket name and path to geojson}

\item{id_col}{the id column for \code{\link[duckdbfs:to_geojson]{duckdbfs::to_geojson()}}}
}
\value{
url to the geojson, e.g. for use with maplibre / mapgl
}
\description{
post as geojson in a public-readable s3 bucket for mapping purposes
}
\details{
Assumes AWS authentication / environmental variables are set.
}
